Release v0.6.0

+++ pp/setting/ppinfo.go
@@ -12,9 +12,9 @@ var IsPP = false
 
 var IsLoginToSP = false
 
-var State byte = ppTypes.PP_INACTIVE
+var State uint32 = ppTypes.PP_INACTIVE
 
-var IsStartMining = false
+var IsStartMining = false // Is the node currently mining
 
 var IsAuto = false
